refactor: #353 - 홈 뷰(운동 중) 애니메이션 Protocol에 의존하도록 변경
